enrollmenttokenapi:Add Region param and WithRegion
This patch modifies all `enrollmenttokenapi` public functions to have a
`Region string` required parameter which is then used to create a new
`context.Context`: `api.WithRegion(context.Background(), params.Region)`
and pass it as the API operation context.

Additionally, all the unit tests have been modified to include a request
assertion which ensures that the region path interpolation is used when
the `/platform` APIs are being called.

Furthermore, the single errors have been removed in favor of prefixed
multi-errors.

Last, the `token.go` and `params.go` file have been split into multiple
files each containing a single API context and its parameters.
