+++ pkg/api/platformapi/enrollmenttokenapi/create.go
@@ -18,14 +18,17 @@
 package enrollmenttokenapi
 
 import (
-	"errors"
+	"context"
 	"fmt"
 	"math"
 	"time"
 
 	"github.com/elastic/cloud-sdk-go/pkg/api"
 	"github.com/elastic/cloud-sdk-go/pkg/api/apierror"
+	"github.com/elastic/cloud-sdk-go/pkg/client/platform_configuration_security"
+	"github.com/elastic/cloud-sdk-go/pkg/models"
 	"github.com/elastic/cloud-sdk-go/pkg/multierror"
+	"github.com/elastic/cloud-sdk-go/pkg/util/ec"
 )
 
 // CreateParams is consumed by Create
@@ -33,58 +36,53 @@ type CreateParams struct {
 	*api.API
 	Roles    []string
 	Duration time.Duration
+	Region   string
 }
 
 // Validate ensures that there's no errors prior to performing the Create API
 // call.
 func (params CreateParams) Validate() error {
-	var merr = multierror.NewPrefixed("enrollment-token create")
+	var merr = multierror.NewPrefixed("invalid enrollment-token create params")
 	if params.API == nil {
 		merr = merr.Append(apierror.ErrMissingAPI)
 	}
 
-	validity := int64(params.Duration.Seconds())
-	if validity > math.MaxInt32 {
+	if validity := int64(params.Duration.Seconds()); validity > math.MaxInt32 {
 		merr = merr.Append(
 			fmt.Errorf("validity value %d exceeds max allowed %d value in seconds", validity, math.MaxInt32),
 		)
 	}
 
-	return merr.ErrorOrNil()
-}
-
-// DeleteParams is consumed by Delete
-type DeleteParams struct {
-	*api.API
-	Token string
-}
-
-// Validate ensures that there's no errors prior to performing the Delete API
-// call.
-func (params DeleteParams) Validate() error {
-	var merr = multierror.NewPrefixed("enrollment-token delete")
-	if params.API == nil {
-		merr = merr.Append(apierror.ErrMissingAPI)
-	}
-
-	if params.Token == "" {
-		merr = merr.Append(errors.New("token cannot be empty"))
+	if err := ec.RequireRegionSet(params.Region); err != nil {
+		merr = merr.Append(err)
 	}
 
 	return merr.ErrorOrNil()
 }
 
-// ListParams is consumed by List
-type ListParams struct {
-	*api.API
-}
-
-// Validate ensures that there's no errors prior to performing the List API
-// call.
-func (params ListParams) Validate() error {
-	if params.API == nil {
-		return apierror.ErrMissingAPI
+// Create creates the token for the specific roles
+func Create(params CreateParams) (*models.RequestEnrollmentTokenReply, error) {
+	if err := params.Validate(); err != nil {
+		return nil, err
 	}
 
-	return nil
+	var persistent = params.Duration.Seconds() <= 0
+	var tokenConfig = models.EnrollmentTokenRequest{
+		Persistent:        ec.Bool(persistent),
+		Roles:             params.Roles,
+		ValidityInSeconds: int32(params.Duration.Seconds()),
+	}
+
+	res, err := params.API.V1API.PlatformConfigurationSecurity.CreateEnrollmentToken(
+		platform_configuration_security.NewCreateEnrollmentTokenParams().
+			WithContext(api.WithRegion(context.Background(), params.Region)).
+			WithBody(&tokenConfig),
+		params.AuthWriter,
+	)
+
+	if err != nil {
+		return nil, api.UnwrapError(err)
+	}
+
+	return res.Payload, nil
 }
